update student room routes

+++ models/user_routes/activity/ActivityRoutes.php
@@ -8,11 +8,15 @@ class ActivityRoutes
     private ActivityService $activity_service;
     private Request $request_data;
     private Middleware $middleware;
+    private $url;
+    private $method;
+    private string $parent_route;
 
-    public function __construct(Request $request_data, Middleware $middleware)
+    public function __construct(Request $request_data, Middleware $middleware, $parent_route)
     {
         $this->request_data = $request_data;
         $this->middleware = $middleware;
+        $this->parent_route = $parent_route;
 
         $this->url = explode('/', $this->request_data->get_request_url());
 
@@ -25,8 +29,8 @@ class ActivityRoutes
 
     public function handle_url()
     {
+        $parent_route = $this->parent_route;
         $request_body = json_decode(file_get_contents('php://input'));
-
         $params = $this->request_data->get_request_params();
         $url = $this->url;
         $count = count($url);
@@ -41,43 +45,46 @@ class ActivityRoutes
             $item_routes = new ItemRoutes($this->request_data, $this->middleware);
             $item_routes->handle_url();
         }else{
+            if ($parent_route == 'instructor') {
 
-            switch ($this->method) {
-                case 'POST':
-                    echo $this->middleware->get_owner_id();
-                    if ($current_route == 'activity' && $count == 1) {
-                        echo json_encode($this->activity_service->add_activity($this->middleware->get_owner_id(), $request_body,));
-                    }
-                    break;
-                case 'GET':
-                    if ($params) {
-                        echo json_encode($this->activity_service->search_activity($params['search']));
-                    }
-
-                    if ($current_route == 'activity' && $count == 2) {
-                        if (intval($next_route) > 0) {
-                            echo json_encode($this->activity_service->get_activity($next_route));
-                        } else {
-                            throw new NotFoundException();
+                switch ($this->method) {
+                    case 'POST':
+                        echo $this->middleware->get_owner_id();
+                        if ($current_route == 'activity' && $count == 1) {
+                            echo json_encode($this->activity_service->add_activity($this->middleware->get_owner_id(), $request_body,));
                         }
-                    } else if ($current_route == 'activities') {
-                        echo json_encode($this->activity_service->get_activities($next_route, $this->middleware->get_owner_id()));
-                    } else {
-                    }
-                    break;
-                case 'PUT':
-                    if ($current_route == 'activity') {
-                        echo json_encode($this->activity_service->update_activity($request_body));
-                    }
-                    break;
-                case 'DELETE':
-                    if ($current_route == 'activity') {
-                        echo json_encode($this->activity_service->delete_activity($next_route));
-                    }
-                    break;
-                default:
-                    throw new NotFoundException();
-                    break;
+                        break;
+                    case 'GET':
+                        if ($params) {
+                            echo json_encode($this->activity_service->search_activity($params['search']));
+                        }
+
+                        if ($current_route == 'activity' && $count == 2) {
+                            if (intval($next_route) > 0) {
+                                echo json_encode($this->activity_service->get_activity($next_route));
+                            } else {
+                                throw new NotFoundException();
+                            }
+                        } else if ($current_route == 'activities') {
+                            echo json_encode($this->activity_service->get_activities($next_route, $this->middleware->get_owner_id()));
+                        }
+                        break;
+                    case 'PUT':
+                        if ($current_route == 'activity') {
+                            echo json_encode($this->activity_service->update_activity($request_body));
+                        }
+                        break;
+                    case 'DELETE':
+                        if ($current_route == 'activity') {
+                            echo json_encode($this->activity_service->delete_activity($next_route));
+                        }
+                        break;
+                    default:
+                        throw new NotFoundException();
+                }
+
+            } else if($parent_route == 'student') {
+                echo 'student activity';
             }
 
         }
